Finish add and remove test Funcions
