+++ src/modules/remove.js
@@ -1,0 +1,7 @@
+import toDoItemsArray from './toDoItemsArray.js';
+
+const removeT = (task) => {
+  toDoItemsArray.splice(task, 1);
+};
+
+export default removeT;
